+++ packages/github-oauth/package.js
@@ -1,13 +1,13 @@
 Package.describe({
   summary: 'GitHub OAuth flow',
-  version: '1.2.0'
+  version: '1.2.1'
 });
 
 Package.onUse(function (api) {
   api.use('oauth2', ['client', 'server']);
   api.use('oauth', ['client', 'server']);
   api.use('http', ['server']);
-  api.use('underscore', 'server');
+  api.use('underscore', ['client', 'server']);
   api.use('random', 'client');
   api.use('service-configuration', ['client', 'server']);