+++ packages/minifier-js/minifier.js
@@ -1,13 +1,13 @@
-var uglify;
+var terser;
 
 meteorJsMinify = function (source) {
   var result = {};
   var NODE_ENV = process.env.NODE_ENV || "development";
 
-  uglify = uglify || Npm.require("uglify-es");
+  terser = terser || Npm.require("terser");
 
   try {
-    var uglifyResult = uglify.minify(source, {
+    var terserResult = terser.minify(source, {
       compress: {
         drop_debugger: false,
         unused: false,
@@ -23,18 +23,18 @@ meteorJsMinify = function (source) {
       }
     });
 
-    if (typeof uglifyResult.code === "string") {
-      result.code = uglifyResult.code;
-      result.minifier = 'uglify-es';
+    if (typeof terserResult.code === "string") {
+      result.code = terserResult.code;
+      result.minifier = 'terser';
     } else {
-      throw uglifyResult.error ||
-        new Error("unknown uglify.minify failure");
+      throw terserResult.error ||
+        new Error("unknown terser.minify failure");
     }
 
   } catch (e) {
     // Although Babel.minify can handle a wider variety of ECMAScript
-    // 2015+ syntax, it is substantially slower than UglifyJS, so we use
-    // it only as a fallback.
+    // 2015+ syntax, it is substantially slower than UglifyJS/terser, so
+    // we use it only as a fallback.
     var options = Babel.getMinifierOptions({
       inlineNodeEnv: NODE_ENV
     });